Improve reusability + Added missing ServiceCollection behavior
