+++ dotnet/src/Connectors/Connectors.MistralAI/MistralAIServiceCollectionExtensions.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System;
+using System.Net.Http;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Microsoft.SemanticKernel.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.MistralAI;
 using Microsoft.SemanticKernel.Embeddings;
@@ -22,18 +24,39 @@ public static class MistralAIServiceCollectionExtensions
     /// <param name="apiKey">The API key required for accessing the Mistral service.</param>
     /// <param name="endpoint">Optional  uri endpoint including the port where MistralAI server is hosted. Default is https://api.mistral.ai.</param>
     /// <param name="serviceId">A local identifier for the given AI service.</param>
+    /// <param name="httpClient">The HttpClient to use with this service.</param>
     /// <returns>The same instance as <paramref name="services"/>.</returns>
     public static IServiceCollection AddMistralChatCompletion(
         this IServiceCollection services,
         string modelId,
         string apiKey,
         Uri? endpoint = null,
-        string? serviceId = null)
+        string? serviceId = null,
+        HttpClient? httpClient = null)
     {
         Verify.NotNull(services);
 
-        return services.AddKeyedSingleton<IChatCompletionService>(serviceId, (serviceProvider, _) =>
-            new MistralAIChatCompletionService(modelId, apiKey, endpoint, HttpClientProvider.GetHttpClient(serviceProvider)));
+        services.AddKeyedSingleton<IChatCompletionService>(serviceId, (serviceProvider, _) =>
+        {
+            var resolvedHttpClient = HttpClientProvider.GetHttpClient(httpClient, serviceProvider);
+
+            if (httpClient == null && serviceProvider?.GetService<HttpClient>() == null)
+            {
+                if (!resolvedHttpClient.DefaultRequestHeaders.Contains("extra-parameters"))
+                {
+                    resolvedHttpClient.DefaultRequestHeaders.Add("extra-parameters", "pass-through");
+                }
+            }
+
+            return new MistralAIChatCompletionService(
+                modelId,
+                apiKey,
+                endpoint,
+                resolvedHttpClient,
+                serviceProvider?.GetService<ILoggerFactory>());
+        });
+
+        return services;
     }
 
     /// <summary>
@@ -44,17 +67,19 @@ public static class MistralAIServiceCollectionExtensions
     /// <param name="apiKey">The API key required for accessing the Mistral service.</param>
     /// <param name="endpoint">Optional  uri endpoint including the port where MistralAI server is hosted. Default is https://api.mistral.ai.</param>
     /// <param name="serviceId">A local identifier for the given AI service.</param>
+    /// <param name="httpClient">The HttpClient to use with this service.</param>
     /// <returns>The same instance as <paramref name="services"/>.</returns>
     public static IServiceCollection AddMistralTextEmbeddingGeneration(
         this IServiceCollection services,
         string modelId,
         string apiKey,
         Uri? endpoint = null,
-        string? serviceId = null)
+        string? serviceId = null,
+        HttpClient? httpClient = null)
     {
         Verify.NotNull(services);
 
         return services.AddKeyedSingleton<ITextEmbeddingGenerationService>(serviceId, (serviceProvider, _) =>
-            new MistralAITextEmbeddingGenerationService(modelId, apiKey, endpoint, HttpClientProvider.GetHttpClient(serviceProvider)));
+            new MistralAITextEmbeddingGenerationService(modelId, apiKey, endpoint, HttpClientProvider.GetHttpClient(httpClient, serviceProvider)));
     }
 }
